fix article img lightbox (#861)
